code reshuffling to avoid import loops between modules
